+++ README.md
@@ -1,6 +1,7 @@
 # Card Game
 
 # 概要
+
 チームで開発を行った、トランプゲームを遊ぶことができるウェブアプリケーションです。</br>
 実装ゲームは、ブラックジャック及びポーカーです。
