+++ package.json
@@ -12,27 +12,25 @@
     "test": "jest"
   },
   "dependencies": {
-    "next": "14.0.2",
+    "next": "^14.0.4",
     "phaser": "^3.70.0",
     "react": "^18",
     "react-dom": "^18"
   },
   "devDependencies": {
     "@babel/core": "^7.23.3",
-    "@babel/preset-env": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
     "@types/jest": "^29.5.10",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
     "@typescript-eslint/parser": "^6.11.0",
-    "autoprefixer": "^10.0.1",
+    "autoprefixer": "^10.4.16",
     "babel-jest": "^29.7.0",
     "eslint": "^8.53.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-next": "14.0.2",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
@@ -42,12 +40,9 @@
     "husky": "^8.0.0",
     "jest": "^29.7.0",
     "lint-staged": "^15.1.0",
-    "postcss": "^8",
     "prettier": "^3.1.0",
-    "tailwindcss": "^3.3.0",
     "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typescript": "5.2.2"
+    "typescript": "^5.3.3"
   },
   "husky": {
     "hooks": {
