+++ src/phaser/blackjack/BlackjackScene.ts
@@ -1,7 +1,7 @@
+import BaseScene from '@/phaser/common/BaseScene'
 import BlackjackTable from '@/models/blackjack/BlackjackTable'
 import TableView from '@/phaser/blackjack/TableView'
 import { GAMETYPE } from '@/types/common/gameTypes'
-import BaseScene from '@/phaser/common/BaseScene'
 
 // PreloadSceneのコードを書き換えることでブラックジャックにできる
 
@@ -19,7 +19,7 @@ export default class BlackjackScene extends BaseScene {
   public create(): void {
     super.create()
 
-    this._tableView = new TableView(this, this._tableModel, GAMETYPE.Blackjack)
+    this._tableView = new TableView(this, this._tableModel)
     // this._tableView.assignDealerBtn()
     // this._tableView.animateCollectBlinds()
