+++ src/phaser/blackjack/PlayerView.ts
@@ -1,4 +1,4 @@
-import Phaser from 'phaser'
+import * as Phaser from 'phaser'
 import Card from '@/models/common/Card'
 import CardView from '@/phaser/common/CardView'
 import BlackjackPlayer from '@/models/blackjack/BlackjackPlayer'
