+++ src/phaser/common/HomeScene.ts
@@ -1,4 +1,4 @@
-import Phaser from 'phaser'
+import * as Phaser from 'phaser'
 
 export default class HomeScene extends Phaser.Scene {
   constructor() {
